prep: ignore jruby alias_method warning
